Fixed a bug in updateTraceBranchConfidence function

+++ src/arch/x86/decoder.hh
@@ -475,7 +475,8 @@ protected:
 
 	// Interface for fetch!
     // tells fetch stage that if a speculative trace is availble for this PC
-	unsigned isTraceAvailable(Addr addr, int64_t value, unsigned confidence);
+    // LVPredictor return int8_t confidence, if this confidence if less than zero then just return
+	unsigned isTraceAvailable(Addr addr, int64_t value, int8_t confidence);
 
     StaticInstPtr getSuperOptimizedMicroop(unsigned traceID, X86ISA::PCState &thisPC, X86ISA::PCState &nextPC, bool &predict_taken);
     void updateStreamTrace(unsigned traceID, X86ISA::PCState &thisPC);
